Split line_buffer module into directory
